+++ packages/gov-ui-kit/src/modules/components/vote/index.ts
@@ -1,2 +1,3 @@
 export * from './voteDataListItem';
 export * from './voteProposalDataListItem';
+export { VoteIndicator } from './voteUtils';